requirements.txt
psycopg2-binary==2.9.9
pandas==2.2.3
numpy==1.26.4
statsmodels==0.14.2
scikit-learn==1.5.0
matplotlib==3.9.0
seaborn==0.13.2
flet==0.25.2
